+++ segment_tree/dual_segment_tree.hpp
