+++ segment_tree/lazy_segment_tree.md
