+++ verify/data_structure/dual_segment_tree_assign.test.cpp
@@ -5,7 +5,7 @@
 #include <limits>
 
 #include "../../algebra/monoid/monoid_assign.hpp"
-#include "../../data_structure/dual_segment_tree.hpp"
+#include "../../segment_tree/dual_segment_tree.hpp"
 
 int main() {
     int N, Q;
