+++ verify/data_structure/dual_segment_tree_plus.test.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 
 #include "../../algebra/monoid/monoid_plus.hpp"
-#include "../../data_structure/dual_segment_tree.hpp"
+#include "../../segment_tree/dual_segment_tree.hpp"
 
 int main() {
     int N, Q;
