+++ verify/data_structure/erasable_priority_queue.test.cpp
@@ -1,8 +1,9 @@
 #define PROBLEM "https://atcoder.jp/contests/abc342/tasks/abc342_g"
 
 #include <iostream>
-#include "data_structure/erasable_priority_queue.hpp"
-#include "data_structure/enumerate_segment_tree_nodes.hpp"
+
+#include "../../data_structure/erasable_priority_queue.hpp"
+#include "../../segment_tree/enumerate_segment_tree_nodes.hpp"
 
 int main() {
     std::ios::sync_with_stdio(false);
@@ -28,17 +29,20 @@ int main() {
         if (type == 1) {
             std::cin >> l[qq] >> r[qq] >> x[qq];
             l[qq]--;
-            auto nodes = enumerate_segment_tree_range_covering_nodes(size, l[qq], r[qq]);
+            auto nodes =
+                enumerate_segment_tree_range_covering_nodes(size, l[qq], r[qq]);
             for (auto&& i : nodes) d[i].push(x[qq]);
         } else if (type == 2) {
             std::cin >> id[qq];
             id[qq]--;
-            auto nodes = enumerate_segment_tree_range_covering_nodes(size, l[id[qq]], r[id[qq]]);
+            auto nodes = enumerate_segment_tree_range_covering_nodes(
+                size, l[id[qq]], r[id[qq]]);
             for (auto&& i : nodes) d[i].erase(x[id[qq]]);
         } else {
             std::cin >> id[qq];
             id[qq]--;
-            auto nodes = enumerate_segment_tree_point_containing_nodes(size, id[qq]);
+            auto nodes =
+                enumerate_segment_tree_point_containing_nodes(size, id[qq]);
             int mx = 0;
             for (auto&& i : nodes) {
                 if (d[i].size() > 0) mx = std::max(mx, d[i].top());
