+++ verify/data_structure/lazy_segment_tree_plus_plus.test.cpp
@@ -4,7 +4,7 @@
 #include <iostream>
 
 #include "../../algebra/acted_monoid/acted_monoid_plus_plus.hpp"
-#include "../../data_structure/lazy_segment_tree.hpp"
+#include "../../segment_tree/lazy_segment_tree.hpp"
 
 int main() {
     int N, Q;
