+++ verify/data_structure/segment_tree_min.test.cpp
@@ -5,7 +5,7 @@
 
 #include "../../algebra/monoid/monoid_max.hpp"
 #include "../../algebra/monoid/monoid_min.hpp"
-#include "../../data_structure/segment_tree.hpp"
+#include "../../segment_tree/segment_tree.hpp"
 
 int main() {
     int N, Q;
