+++ verify/graph/heavy_light_decomposition_vertex_add_subtree_sum.test.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 
 #include "../../algebra/monoid/monoid_plus.hpp"
-#include "../../data_structure/segment_tree.hpp"
+#include "../../segment_tree/segment_tree.hpp"
 #include "../../graph/heavy_light_decomposition.hpp"
 #include "../../graph/read_graph.hpp"
